requirements.txt
python-telegram-bot==13.15
Flask==2.3.3
openai==1.14.2
